move lazy utils to modules.lua
